+++ packages/reactive-element/src/decorators/base.ts
@@ -4,30 +4,17 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
-import {ReactiveElement} from '../reactive-element.js';
-
-export type Constructor<T> = {
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  new (...args: any[]): T;
-};
-
-// From the TC39 Decorators proposal
-export interface ClassDescriptor {
-  kind: 'class';
-  elements: ClassElement[];
-  finisher?: <T>(clazz: Constructor<T>) => void | Constructor<T>;
-}
-
-// From the TC39 Decorators proposal
-export interface ClassElement {
-  kind: 'field' | 'method';
-  key: PropertyKey;
-  placement: 'static' | 'prototype' | 'own';
-  initializer?: Function;
-  extras?: ClassElement[];
-  finisher?: <T>(clazz: Constructor<T>) => void | Constructor<T>;
-  descriptor?: PropertyDescriptor;
-}
+import {
+  ReactiveElement,
+  ClassElement,
+  Constructor,
+} from '../reactive-element.js';
+export {
+  Constructor,
+  ClassElement,
+  ClassDescriptor,
+  CompatiblePropertyDecorator,
+} from '../reactive-element.js';
 
 export const legacyPrototypeMethod = (
   descriptor: PropertyDescriptor,